Refactor React APP + properties + proxy
