+++ sources/platform/server/src/main/frontend/src/components/SearchBar/index.js
